Separated scheduler
