+++ image_generation/stable_diffusion_1_5/cpp/src/main.cpp
@@ -7,50 +7,21 @@
 int32_t main(int32_t argc, char* argv[]) {
     cxxopts::Options options("OV_SD_CPP", "SD implementation in C++ using OpenVINO\n");
 
-    options.add_options()(
-        "p,posPrompt",
-        "Initial positive prompt for SD ",
-        cxxopts::value<std::string>()->default_value(
-        "cyberpunk cityscape like Tokyo New York  with tall buildings at dusk golden hour cinematic lighting"))(
-        "n,negPrompt",
-        "Defaut is empty with space",
-        cxxopts::value<std::string>()->default_value(" "))(
-        "d,device",
-        "AUTO, CPU, or GPU",
-        cxxopts::value<std::string>()->default_value("CPU"))(
-        "step",
-        "Number of diffusion steps",
-        cxxopts::value<size_t>()->default_value("20"))(
-        "s,seed",
-        "Number of random seed to generate latent for one image output",
-        cxxopts::value<size_t>()->default_value("42"))(
-        "num",
-        "Number of image output",
-        cxxopts::value<size_t>()->default_value("1"))(
-        "height",
-        "height",
-        cxxopts::value<size_t>()->default_value("512"))(
-        "width",
-        "width",
-        cxxopts::value<size_t>()->default_value("512"))(
-        "c,useCache",
-        "use model caching",
-        cxxopts::value<bool>()->default_value("false"))(
-        "r,readNPLatent",
-        "read numpy generated latents from file",
-        cxxopts::value<bool>()->default_value("false"))(
-        "m,modelPath",
-        "Specify path of SD model IR",
-        cxxopts::value<std::string>()->default_value("../models/dreamlike-anime-1.0"))(
-        "t,type",
-        "Specify the type of SD model IR(FP16_static or FP16_dyn)",
-        cxxopts::value<std::string>()->default_value("FP16_static"))(
-        "l,loraPath",
-        "Specify path of lora file. (*.safetensors).",
-        cxxopts::value<std::string>()->default_value(""))(
-        "a,alpha",
-        "alpha for lora",
-        cxxopts::value<float>()->default_value("0.75"))("h,help", "Print usage");
+    options.add_options()
+    ("p,posPrompt", "Initial positive prompt for SD ", cxxopts::value<std::string>()->default_value("cyberpunk cityscape like Tokyo New York  with tall buildings at dusk golden hour cinematic lighting"))
+    ("n,negPrompt","Defaut is empty with space", cxxopts::value<std::string>()->default_value(" "))
+    ("d,device", "AUTO, CPU, or GPU", cxxopts::value<std::string>()->default_value("CPU"))
+    ("step", "Number of diffusion steps", cxxopts::value<size_t>()->default_value("20"))
+    ("s,seed", "Number of random seed to generate latent for one image output", cxxopts::value<size_t>()->default_value("42"))
+    ("num", "Number of image output", cxxopts::value<size_t>()->default_value("1"))
+    ("height", "destination image height", cxxopts::value<size_t>()->default_value("512"))
+    ("width", "destination image width", cxxopts::value<size_t>()->default_value("512"))
+    ("c,useCache", "use model caching", cxxopts::value<bool>()->default_value("false"))
+    ("r,readNPLatent", "read numpy generated latents from file", cxxopts::value<bool>()->default_value("false"))
+    ("m,modelPath", "Specify path of SD model IRs", cxxopts::value<std::string>()->default_value("../models/dreamlike-anime-1.0"))
+    ("t,type", "Specify the type of SD model IRs (e.g., FP16_static or FP16_dyn)", cxxopts::value<std::string>()->default_value("FP16_static"))
+    ("l,loraPath", "Specify path of LoRA file. (*.safetensors).", cxxopts::value<std::string>()->default_value(""))
+    ("a,alpha", "alpha for LoRA", cxxopts::value<float>()->default_value("0.75"))("h,help", "Print usage");
     cxxopts::ParseResult result;
 
     try {
